+++ CHANGELOG.md
@@ -11,10 +11,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `PortDiff::resolve_boundary_index` to handle wire boundaries.
 - Add `PortDiffGraph::from_sinks_while` to create a `PortDiffGraph` from a predicate.
 - Add `PortDiffGraph::is_squashable` and `PortDiffGraph::try_squash` to check and squash compatible `PortDiffGraph`s.
+- Add `PortDiff::all_parents` to get all parent `PortDiff`s.
+- Viewer: show CX count delta in hierarchy node labels.
 
 ### Changed
 - Redesigned boundary port abstraction. Use `BoundarySite::Wire` to represent a boundary not connected to a site.
 - Rename `GraphView` to `PortDiffGraph` to better reflect its purpose.
+
+### Fixed
+- Crash on squashing several wires with identical IDs.
 
 ## [0.2.0] - 2024-09-03